Extract out NTAuthentication to use in System.Net.Mail. Restructure System.Net.Security to consume common components.
