+++ src/Common/src/System/Net/ContextFlagsAdapterPal.Windows.cs
