+++ src/Common/src/System/Net/NegotiationInfoClass.cs
