+++ src/Common/src/System/Net/Security/SSPIHandleCache.cs
