+++ src/Common/src/System/Net/Security/SecurityBuffer.cs
