+++ src/Common/src/System/Net/Security/SecurityBufferType.cs
