+++ src/Common/src/System/Net/Security/SecurityContextTokenHandle.cs
