+++ src/Common/src/System/Net/SecurityStatusAdapterPal.Windows.cs
